requirements.txt
numpy
opencv-python
mtcnn
scikit-learn
playsound==1.2.2
Pillow
tensorflow-gpu